Fix to properly handle spaces in arguments
Command like tfe pushvars -hcl-var 'h=["one", "two"]' were failing. The space inside the HCL variable caused the argument to be split due to the way the arguments were being assembled and passed to subcommands. The fix is to leave the original $@ along, passing it forward and providing a way to detect and skip common arguments that have already been handled.